Allow Add-Existing-Repository to validate home/tilde paths
There already exists a function that will convert a tilde path to an absolute path. It was originally used for this purpose, but the functionality was removed duringa commit that changed which function was used to validate git repositories.

This reinstates that functionality and allows us to type a ~/ tilde path to get our home directories when typing in a path.